LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/request-headers-expected.txt


FAIL Fetch with GET assert_equals: Request should have header origin: http://localhost:8800 expected (string) "http://localhost:8800" but got (object) null
FAIL Fetch with HEAD assert_equals: Request should have header content-length: null expected (object) null but got (string) "0"
PASS Fetch with PUT without body 
PASS Fetch with PUT with body 
PASS Fetch with POST without body 
PASS Fetch with POST with text body 
FAIL Fetch with POST with FormData body promise_test: Unhandled rejection with value: object "TypeError: Uploading FormData is not yet implemented"
FAIL Fetch with POST with Blob body assert_equals: Request should have header content-type: null expected (object) null but got (string) "application/x-www-form-urlencoded"
FAIL Fetch with POST with ArrayBuffer body promise_test: Unhandled rejection with value: object "TypeError: Type error"
PASS Fetch with POST with Blob body with mime type 
FAIL Fetch with Chicken assert_equals: Request should have header content-length: null expected (object) null but got (string) "0"
PASS Fetch with Chicken with body 
